DJANGO/requirements.txt
argparse
webb
dnspython
requests
python-whois
sublist3r
rfc3987
termcolor
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.2
Django==3.2.5
dnspython==2.1.0
future==0.18.2
idna==3.2
python-whois==0.7.3
pytz==2021.1
requests==2.26.0
rfc3987==1.3.8
sqlparse==0.4.1
Sublist3r==1.0
termcolor==1.1.0
urllib3==1.26.6
webb==0.9.2.5